WEB_PAGE/requirements.txt
flask
numpy
opencv-python
deepface
google-generativeai
tf-keras